Begin Work on refactoring. Added Notes on DP init. Code does not work in current state. Need to re-write controller start to incorporate moved functions. Need to delegate more processing to these new functions
