+++ i915_reg.h
@@ -1,0 +1,2 @@
+#define PCH_DISPLAY_BASE	0xc0000u
+#define DETAIL_TIME_SELCTION 1
